[Refactor] Move addUefaTokenIfAny to ImportToken provider, Remove unused dependencies #4953
