+++ AlphaWallet/Core/TokensAutodetector/ImportToken.swift
@@ -18,6 +18,7 @@ class ImportToken {
     private let analyticsCoordinator: AnalyticsCoordinator
     private let tokenFetchers: AtomicDictionary<RPCServer, TokenFetcher> = .init()
     private let tokensDataStore: TokensDataStore
+    private var cancelable = Set<AnyCancellable>()
 
     let wallet: Wallet
 
@@ -27,6 +28,22 @@ class ImportToken {
         self.assetDefinitionStore = assetDefinitionStore
         self.analyticsCoordinator = analyticsCoordinator
         self.wallet = wallet
+
+        addUefaTokenIfAny()
+    }
+
+    private func addUefaTokenIfAny() {
+        guard !isRunningTests() else { return }
+        
+        //NOTE: initally when we set sessions, we want to import uefa tokens, for enabled chain
+        sessions.filter { !$0.values.isEmpty }
+            .first()
+            .sink { _ in
+                let server = Constants.uefaRpcServer
+                self.importToken(for: Constants.uefaMainnet, server: server, onlyIfThereIsABalance: true)
+                    .done { _ in }
+                    .cauterize()
+            }.store(in: &cancelable)
     }
 
     //Adding a token may fail if we lose connectivity while fetching the contract details (e.g. name and balance). So we remove the contract from the hidden list (if it was there) so that the app has the chance to add it automatically upon auto detection at startup
